seperated character_model from main
seperated character_model from main
