+++ main.py
@@ -6,12 +6,13 @@ from jsonschema import validate
 from schemas import character_model_schema
 from model_to_gui import convert_model_to_gui
 from gui_to_model import convert_gui_to_model
+import character_model
 import json, csv
 
 class Character:
     def __init__(self, root) -> None:
         self.root = root
-        self.character_model = Character_Model()
+        self.character_model = character_model.Character_Model()
 
         self.character_class_option = StringVar(root)
         self.character_race_option = StringVar(root)
@@ -582,9 +583,6 @@ class Character:
                 else:
                     skill[1].set(False)
 
-
-
-
     def determine_modifier(self, score):
         if score % 2 != 0:
             score -= 1
@@ -737,87 +735,6 @@ class Character:
         except Exception as e:
             messagebox.showerror("An error occurred", str(e))
 
-
-class Character_Model:
-    def __init__(self) -> None:
-        self.character_name = ""
-        self.character_class = ""
-        self.race = ""
-        self.level = 0
-        self.attributes = {
-            'strength': 0,
-            'dexterity': 0,
-            'constitution': 0,
-            'intelligence': 0,
-            'wisdom': 0,
-            'charisma': 0
-        }
-        self.hitpoints = {
-            'max': 0,
-            'current': 0
-        }
-        self.armor_class = 0
-        self.initiative = 0
-        self.speed = 0
-        self.saving_throws = {
-            'strength': False,
-            'dexterity': False,
-            'constitution': False,
-            'intelligence': False,
-            'wisdom': False,
-            'charisma': False
-        }
-        self.hit_dice = {
-            'total': "",
-            'current': ""
-        }
-        self.death_saves = {
-            'successes': [False, False, False],
-            'failures': [False, False, False]
-        }
-        # Weapon attacks will be in a form of a list of dictionaries
-        # Example: [{'name': 'Sword', 'bonus_to_hit': '5', 'damage': '1d6 + 3'}, ...]
-        self.weapon_attacks = []
-        self.abilities = ""
-        self.skills = {
-            'acrobatics': False,
-            'animal_handling': False,
-            'arcana': False,
-            'athletics': False,
-            'deception': False,
-            'history': False,
-            'insight': False,
-            'intimidation': False,
-            'investigation': False,
-            'medicine': False,
-            'nature': False,
-            'perception': False,
-            'performance': False,
-            'persuasion': False,
-            'religion': False,
-            'sleight_of_hand': False,
-            'stealth': False,
-            'survival': False
-        }
-        self.other_proficiencies = ""
-        self.inventory = ""
-        self.spells = ""
-
-    # Saving Data (instance attributes) to a JSON file. 
-    # This will be used to save the Character Model
-    def to_json(self):
-        return json.dumps(self.__dict__, indent=4)
-
-    @classmethod
-    def from_json(cls, json_str):
-        try:
-            instance = cls()
-            instance.__dict__ = json.loads(json_str)
-            return instance
-        except json.JSONDecodeError as e:
-            print(f"Invalid JSON: {e}")
-            return None
-
 if __name__ == '__main__':
     root = Tk()
     root.title('Character Smith')
@@ -825,18 +742,6 @@ if __name__ == '__main__':
     root.bind_all("<Button-1>", lambda event: event.widget.focus_set())
 
     character = Character(root)
-
-    # JSON FILE TEST
-    # with open('character.json', 'w') as f:
-    #     f.write(character.character_model.to_json())
-    # 
-    # with open('character.json', 'r') as f:
-    #     json_str = f.read()
-    #     loaded_character = character.character_model.from_json(json_str)
-    #     # TODO: Prompt user before loading the character_model
-    #     character.character_model = loaded_character
-    # 
-    # print("Level: " + str(character.character_model.level))
 
     print('Running...')
